Added discovery package. Updated docs for config discovery.
