+++ docs/book/SUMMARY.md
@@ -119,6 +119,11 @@
   * [Service API](decision_logs.md#decision-log-service-api)
 * [Monitoring](monitoring.md)
   * [Prometheus](monitoring.md#prometheus)
+* [Discovery](discovery.md)
+  * [Service API](discovery.md#discovery-service-api)
+* [Monitoring & Diagnostics](monitoring-diagnostics.md)
+  * [Prometheus](monitoring-diagnostics.md#prometheus)
+  * [Diagnostics](monitoring-diagnostics.md#prometheus)
 * [Security](security.md)
   * [TLS and HTTPS](security.md#tls-and-https)
   * [Authentication and Authorization](security.md#authentication-and-authorization)
